添加meta信息处理

+++ SM/SM_Manager.cpp
@@ -3,9 +3,19 @@
 //
 
 #include "SM_Manager.h"
+#include "../utils/Utils.h"
 
-SM_Manager::SM_Manager() : ixManager(IX_Manager::Instance()), rmManager(RM_Manager::Instance()), isUsingDb(false) {
+SM_Manager::SM_Manager() : pfManager(PF_Manager::Instance()), rmManager(RM_Manager::Instance()),
+                           ixManager(IX_Manager::Instance()), isUsingDb(false) {
 
+    //Meta文件中存储了所有的数据库管理系统的信息
+    dbmsMetaFile = Utils::getAbsolutePath("DBMS.Meta");
+    //如果没有该文件，那么要初始化一个带meta的meta文件，否则从该文件中读取meta
+    if (access(dbmsMetaFile.c_str(), F_OK) < 0) {
+        pfManager.CreateFileWithMeta(dbmsMetaFile.c_str(), &dbmsMeta, sizeof(DbmsMeta));
+    } else {
+        pfManager.GetMeta(dbmsMetaFile.c_str(), &dbmsMeta, sizeof(DbmsMeta));
+    }
 }
 
 SM_Manager::~SM_Manager() {
@@ -60,17 +70,34 @@ RC SM_Manager::Set(const char *paramName, const char *value) {
 }
 
 RC SM_Manager::CreateDb(const char *dbName) {
-    if (mkdir(dbName, S_IRWXU) < 0) {
-        return SM_DB_ALREADY_IN;
+    for (auto &db : dbmsMeta.databaseName)
+        if (strcmp(db, dbName) == 0) {
+            return SM_DB_ALREADY_IN;
+        }
+    for (auto &i : dbmsMeta.databaseName) {
+        if (strlen(i) == 0) {
+            strcpy(i, dbName);
+            if (mkdir(dbName, S_IRWXU) < 0) {
+                return SM_UNIX;
+            }
+            TRY(pfManager.UpdateMeta(dbmsMetaFile.c_str(), &dbmsMeta, sizeof(DbmsMeta)));
+            return OK_RC;
+        }
     }
-    return OK_RC;
+    return SM_DB_NUM_OVER;
 }
 
 RC SM_Manager::DropDb(const char *dbName) {
-    if (rmdir(dbName) < 0) {
-        return SM_DB_NOT_EXIST;
-    }
-    return OK_RC;
+    for (auto &db : dbmsMeta.databaseName)
+        if (strcmp(db, dbName) == 0) {
+            memset(db, 0, sizeof(db));
+            if (rmdir(dbName) < 0) {
+                return SM_UNIX;
+            }
+            TRY(pfManager.UpdateMeta(dbmsMetaFile.c_str(), &dbmsMeta, sizeof(DbmsMeta)));
+            return OK_RC;
+        }
+    return SM_DB_NOT_EXIST;
 }
 
 RC SM_Manager::UseDb(const char *dbName) {
@@ -93,4 +120,31 @@ RC SM_Manager::DropColumn(const char *dbName, const char *columnName) {
 SM_Manager &SM_Manager::Instance() {
     static SM_Manager instance;
     return instance;
+}
+
+RC SM_Manager::ShowDatabases() {
+    printf("this is me!!!!\n");
+    int cnt = 0;
+    for (auto &db : dbmsMeta.databaseName)
+        if (strlen(db) > 0) {
+            cnt++;
+            printf("%20s", db);
+            if (++cnt % 5 == 0) {
+                printf("\n");
+            }
+        }
+    printf("\n");
+    return OK_RC;
+}
+
+RC SM_Manager::DescDatabase(const char &dbName) {
+    return PF_EOF;
+}
+
+RC SM_Manager::ShowTables() {
+    return PF_EOF;
+}
+
+RC SM_Manager::DescTable(const char *tbName) {
+    return PF_EOF;
 }
